Split cuprum/context.py into a context/ package (#116)
`cuprum/context.py` exceeded the 400-line ceiling mandated by AGENTS.md
and mixed four concerns that change for different reasons: pure
env-overlay merging, the core domain dataclasses, the ContextVar
plumbing, and the user-facing registration handles.

Split into a `cuprum/context/` package, re-exporting the existing public
`__all__` unchanged from `__init__.py` so importers are unaffected:

- `context/env_overlay.py` — `_coerce_env_overlay`, `merge_env_overlays`,
  `resolve_env`; pure overlay logic with no ContextVar dependency.
- `context/core.py` — `ForbiddenProgramError`, `_validate_timeout`,
  `ScopeConfig`, `CuprumContext`, and the hook type aliases.
- `context/state.py` — the ContextVar machinery (`current_context`,
  `get_context`, internal set/reset helpers).
- `context/registration.py` — `_ScopedContext`, `scoped`, the
  `_TokenRegistration` base, the registration handles, and the
  `allow`/`before`/`after`/`env`/`observe` factories, landing already
  de-duplicated on top of the #113 extraction.

Every resulting module is under 400 lines (the largest is 365). The code
is moved verbatim apart from import plumbing; the existing public-API,
context, env-overlay, and stateful registration suites pass unchanged.
Update `docs/developers-guide.md` with the package layout and refresh the
wheel-build snapshot for the new file list.
